Bugfix: zero elements in nodeSegN
If the data contained single nodes that were not connected to any edges, then the list nodeEdges would not contain the indices for these nodes. Instead, it would contain a value of 0 from when the array was initialized.

When down sampling (Regraph Nodes),  the script nodeGrps_vesSegment would try to use this zero element as an index, which would raise an error.

The bugfix was to remove these floating nodes in the first steps of "Get Segment Info > Update"