Develop (#4)
* Function to remove islands of pixels

* Updated file paths/ma,es for island removed

* Corrected filename loading error.

* Updated filename out. Removed redundancy

* Created function for segment to graph.

Need to test with main script.

* Added comments

* Added test script for debugging zero index error in nodeGrps_vesSegment

This error is raising an error when trying to delete segments in the GUI.

* Function: remove floating nodes & reindex edges

* Examine output of nodeGrps_vesSegment test script

* Removed outdated fig

* Bugfix: zero elements in nodeSegN

If the data contained single nodes that were not connected to any edges, then the list nodeEdges would not contain the indices for these nodes. Instead, it would contain a value of 0 from when the array was initialized.

When down sampling (Regraph Nodes),  the script nodeGrps_vesSegment would try to use this zero element as an index, which would raise an error.

The bugfix was to remove these floating nodes in the first steps of "Get Segment Info > Update"

* Changed filename - Stephan new smoothing

* Fixed filename extension for new smoothing/segmentation

* Updated output file type ('.tif' not '.btf')

* Updated main script and "seg_to_graph" function

Script calls segmentation and segment->graph function.
Need to add initialization of GUI.

* Updated comment for adding path

* Main script: Added step to mask vessel segments

Need to:
- create function to create binary mask from 3D volume
- create function for applying mask

* Fixed call to unverifiedIdx

Now, we can trash deleted segments.

However, the variables "segmentsGrpOrder" and "segmentsUnverifiedGrpOrder" need to be updated to the correct size. They are not yet being updated.

* Added SCC directory for testing latest data

* Fix segCgrps and grporder issue when segments are trashed

* Added comments to function for deleting segments

Need to pull in changes from Sreekanth

* Bugfix: select node with multiple segments

The code was breaking when trying to select an end node. It was being handled as a node with multiple segments. This was resolved by replacing "<" with "<="

* Updated variable names in test scripts

Tested workflow before merging pull request

---------

Co-authored-by: sreekanthkura7 <sreekanth.kura7@gmail.com>