Root cause of bug in downsample_segment_group
The cell array of group_node_idcs contained duplicate node indices across multiple groups (arrays). This led to a mismatch between the total number of nodes and the largest edge index, which caused the remapped edges to be jumbled.

The bugfix will be to remove these duplicate nodes from the groups