Identified end points of non-loop segments
The end points of the non-loop segments should be maintained during down sampling. This will preserve the morphology of the vasculature while removing the loops. These points were identified and they were added to "end_nodes" array, which are defined as unique nodes in the down sampling for loop.

TODO:
1)