Found bug. Fix proposed in comments of script
In the outer while-loop, I select a subset of nodes to downsample. This subset is the first row in the cell array "cnodes". However, in the case that there are nested/connected loops, these nodes will be contained within multiple cell array rows. Therefore, the code is currently unable to process nested/connected loops because the variable "n_idcs" does not contain all the nodes within the nested structure. This is raising an error later in the code when trying to graph because nodes are being excluded.