30 parallelize loop removal (#31)
The purpose of this merge was to include the updated "rm_loops" function. This updated function includes a boolean input argument (viz) that determines whether to visualize figures. This should be set to false when removing loops from large datasets.

* Added test script for parallelizing loop removal

- added TODO at top

* removed deprecated script

This was created before I joined. However, it is deprecated.

* input argument - enable visualize figures

The input argument "viz" is a boolean that determines whether to display figures. This is useful for debugging, but it computationally expensive and should be set to false for large datasets.