Corrected graph call and denominator for VF, LD, BD
I was importing the incorrect graph when calculating the average value for each metric. I corrected this and reran the statistical tests. The result was a lack of significance (p<0.05) for all metrics.

For the heatmaps, I was incorrectly calculating the denominator for the volume fraction, length density, and branch density. I was using the volume of an ROI as the denominator. Instead, I should have been using the volume of tissue voxels (inclusive of vasculature) as the denominator. I corrected this mistake and reran the statistical tests.